requirements.txt
PyQt6==6.6.1
PyQt6-WebEngine==6.6.0
Pillow==10.2.0
openai==1.12.0 